requirements.txt
openai>=1.12.0
google-genai==1.56.60
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.32.0
